pyproject.toml
[tool.poetry]
name = "soil-fertility"
version = "0.1.0"
description = ""
authors = ["RedhaWassim <rwasssim@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.9.7 || >3.9.7,<3.11"
pandas = "^2.1.3"
mypy = "^1.7.0"
black = "^23.11.0"
ruff = "^0.1.5"
codespell = "^2.2.6"
pytest = "^7.4.3"
ipykernel = "^6.26.0"
matplotlib = "^3.8.1"
seaborn = "^0.13.0"
openpyxl = "^3.1.2"
scikit-learn = "^1.3.2"
pydantic = "^2.5.1"
streamlit = "^1.29.0"
pyyaml = "^6.0.1"
plotly = "^5.18.0"


[tool.poetry.group.stubs.dependencies]
pandas-stubs = "^2.1.1.230928"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

